Fix a bunch of warnings (Reduce use of bitfields)
Fix the many warnings that existed, from unused variables to actual
errors. All the while, the use of bitfields was reduced a lot, mostly
because using them was creating complications for not a lot of gain.

That is, to use them without warning we'd need either:
  data->option = (TypeCast) value & MAX_VALUE;
or even (because gboolean is signed) :
  data->is_flag = (guint) value & 1;

Bitfields are still used for internal flags or in node/task as those are
quite common/used a lot. But for treeviews & columntypes it might not be
worth the trouble.

+++ src/history.c
@@ -1,4 +1,6 @@
 
+#include "config.h"
+
 #include <string.h>     /* memmove() */
 #include "history.h"
 #include "macros.h"
@@ -47,7 +49,7 @@ donna_history_set_max (DonnaHistory           *history,
     }
 
     history->max = max;
-    g_ptr_array_set_size (history->items, max);
+    g_ptr_array_set_size (history->items, (gint) max);
 }
 
 guint
